+++ src/lib/hooks/useOpenTab.ts
@@ -18,11 +18,11 @@ export const useOpenTxTab = (type: "lcd" | "tx-page") => {
 };
 
 export const useOpenAssetTab = () => {
-  const balancesApiRoute = useBaseApiRoute("balances");
+  const balancesApiRoute = useBaseApiRoute("accounts");
 
   return useCallback(
     (walletAddr: string) => {
-      openNewTab(`${balancesApiRoute}/${walletAddr}`);
+      openNewTab(`${balancesApiRoute}/${walletAddr}/balances`);
     },
     [balancesApiRoute]
   );