+++ src/lib/pages/contract-details/components/tables/TxsTable.tsx
@@ -1,22 +1,20 @@
-import type { ChangeEvent } from "react";
-
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
 import { EmptyState } from "lib/components/state";
 import { TransactionsTable } from "lib/components/table";
 import { DEFAULT_TX_FILTERS } from "lib/data";
-import { useTxsByAddressPagination } from "lib/services/txService";
-import type { Nullable, Option } from "lib/types";
+import { useTxsByAddress } from "lib/services/txService";
+import type { ContractAddr, Option } from "lib/types";
 
 interface TxsTableProps {
-  contractAccountId: Option<Nullable<number>>;
+  contractAddress: ContractAddr;
   scrollComponentId: string;
   totalData: Option<number>;
   refetchCount: () => void;
 }
 
 export const TxsTable = ({
-  contractAccountId,
+  contractAddress,
   scrollComponentId,
   totalData,
   refetchCount,
@@ -37,31 +35,19 @@ export const TxsTable = ({
     },
   });
 
-  const { data: transactions, isLoading } = useTxsByAddressPagination(
-    contractAccountId,
-    "",
-    DEFAULT_TX_FILTERS,
+  const { data, isLoading } = useTxsByAddress(
+    contractAddress,
     undefined,
+    undefined,
+    DEFAULT_TX_FILTERS,
     offset,
     pageSize
   );
 
-  const onPageChange = (nextPage: number) => {
-    refetchCount();
-    setCurrentPage(nextPage);
-  };
-
-  const onPageSizeChange = (e: ChangeEvent<HTMLSelectElement>) => {
-    const size = Number(e.target.value);
-    refetchCount();
-    setPageSize(size);
-    setCurrentPage(1);
-  };
-
   return (
     <>
       <TransactionsTable
-        transactions={transactions}
+        transactions={data?.items}
         isLoading={isLoading}
         emptyState={
           <EmptyState
@@ -79,8 +65,16 @@ export const TxsTable = ({
           totalData={totalData}
           scrollComponentId={scrollComponentId}
           pageSize={pageSize}
-          onPageChange={onPageChange}
-          onPageSizeChange={onPageSizeChange}
+          onPageChange={(nextPage) => {
+            setCurrentPage(nextPage);
+            refetchCount();
+          }}
+          onPageSizeChange={(e) => {
+            const size = Number(e.target.value);
+            setPageSize(size);
+            setCurrentPage(1);
+            refetchCount();
+          }}
         />
       )}
     </>